feat: Add popDomainEvents method to AggregateRoot
This commit adds a new method, popDomainEvents, to the AggregateRoot class. This method retrieves and clears the list of domain events stored in the AggregateRoot. The method is then used in the Mediator class to publish the domain events to the pipeline. This change is a new feature for the user and will trigger a release bumping a MINOR version.